requirement.txt
django
django-sesame
